requirements.txt
sqlalchemy
pandas
python-dotenv
ollama
requests
